refactor: move StatsTracer to stats/rtc
